Issue #46 - Change PostId type to long in ForumPost table
Updated the `PostId` column type from `int` to `long` in the `ForumPost` table creation within the migration file `20241029184053_AddForumTables.cs`. This modification allows for a larger range of values, accommodating more posts in the forum.

+++ source/digioz.Forum/digioz.Forum/Data/Migrations/20241029184053_AddForumTables.cs
@@ -374,7 +374,7 @@ namespace digioz.Forum.Data.Migrations
                 name: "ForumPost",
                 columns: table => new
                 {
-                    PostId = table.Column<int>(type: "int", nullable: false)
+                    PostId = table.Column<long>(type: "int", nullable: false)
                         .Annotation("SqlServer:Identity", "1, 1"),
                     TopicId = table.Column<int>(type: "int", nullable: false, defaultValue: 0),
                     ForumId = table.Column<long>(type: "bigint", nullable: false, defaultValue: 0),
